Banner, Input Tabun

+++ app/Http/Controllers/ArticleShowController.php
@@ -60,9 +60,7 @@ class ArticleShowController extends Controller
             $validated = $request->validate([
                 'judul' => 'required|max:255|unique:' . ArticleShow::class,
                 'category' => 'array',
-                'category.*' => 'max:255',
                 'tag' => 'array',
-                'tag.*' => 'max:255',
                 'article' => 'required',
             ]);
     
@@ -96,186 +94,186 @@ class ArticleShowController extends Controller
                 ->withErrors($e->validator);
         }
 
-        DB::transaction(function () use ($request) {
-            // Article
-            $newarticle = new Article;
-    
-            $newarticle->user_id = Auth::id();
-            $newarticle->judul = $request->judul;
-            $newarticle->article = $request->article;
-    
-            if ($request->status === "schedule") {
-                $newarticle->schedule = true;
-            } else {
-                $newarticle->schedule = false;
-            }
-    
-            // Cek apakah link adalah YouTube
-            if (preg_match('/(?:youtube\.com\/(?:[^\/]+\/.+\/|(?:v|e(?:mbed)?)\/|.*[?&]v=)|youtu\.be\/)([^"&?\/\s]{11})/', $request->link, $matches)) {
-                $videoId = $matches[1];
-                $newarticle->video_type = "youtube";
-                $newarticle->youtube = "https://www.youtube.com/embed/{$videoId}";
-                $newarticle->tiktok = null;
-            } elseif (preg_match('/(?:www\.)?tiktok\.com\/(@[\w.-]+)\/video\/(\d+)/', $request->link, $matches)) {
-                $newarticle->video_type = "tiktok";
-                $newarticle->youtube = null;
-                $newarticle->tiktok = "https://www.tiktok.com/{$matches[0]}";
-            } else {
-                $newarticle->video_type = "none";
-                $newarticle->youtube = null;
-                $newarticle->tiktok = null;
-            }
-    
-            $newarticle->save();
-    
-            $newarticle->template()->sync([$request->template_id]);
-            
-            $newbanner = null;
-    
-            // Banner
-            if ($request->hasFile('image')) {
-                $newbanner = new ArticleBanner;
+        // Article
+        $newarticle = new Article;
+
+        $newarticle->user_id = Auth::id();
+        $newarticle->judul = $request->judul;
+        $newarticle->article = $request->article;
+
+        if ($request->status === "schedule") {
+            $newarticle->schedule = true;
+        } else {
+            $newarticle->schedule = false;
+        }
+
+        // Cek apakah link adalah YouTube
+        if (preg_match('/(?:youtube\.com\/(?:[^\/]+\/.+\/|(?:v|e(?:mbed)?)\/|.*[?&]v=)|youtu\.be\/)([^"&?\/\s]{11})/', $request->link, $matches)) {
+            $videoId = $matches[1];
+            $newarticle->video_type = "youtube";
+            $newarticle->youtube = "https://www.youtube.com/embed/{$videoId}";
+            $newarticle->tiktok = null;
+        } elseif (preg_match('/(?:www\.)?tiktok\.com\/(@[\w.-]+)\/video\/(\d+)/', $request->link, $matches)) {
+            $newarticle->video_type = "tiktok";
+            $newarticle->youtube = null;
+            $newarticle->tiktok = "https://www.tiktok.com/{$matches[0]}";
+        } else {
+            $newarticle->video_type = "none";
+            $newarticle->youtube = null;
+            $newarticle->tiktok = null;
+        }
+
+        $newarticle->save();
+
+        $newarticle->template()->sync([$request->template_id]);
         
-                $newbanner->article_id = $newarticle->id;
+        $newbanner = null;
+
+        // Banner
+        if ($request->hasFile('image')) {
+            $newbanner = new ArticleBanner;
     
-                $imageFile = $request->file('image');
-                $imageName = time();
-                $imagePath = public_path('storage/images/article/banner/');
-    
-                // Pastikan direktori ada, jika tidak maka buat
-                if (!File::exists($imagePath)) {
-                    File::makeDirectory($imagePath, 0755, true);
-                }
-    
-                $manager = new ImageManager(new Driver());
-                $image = $manager->read($imageFile->getPathname());
-    
-                $imageFullPath = $imagePath . $imageName . '.webp';
-                $image->save($imageFullPath);
-    
-                $newbanner->image = $imageName . '.webp';
-                $newbanner->image_alt = $imageName;
-                
-                $newbanner->save();
+            $newbanner->article_id = $newarticle->id;
+
+            $imageFile = $request->file('image');
+            $imageName = time();
+            $imagePath = public_path('storage/images/article/banner/');
+
+            // Pastikan direktori ada, jika tidak maka buat
+            if (!File::exists($imagePath)) {
+                File::makeDirectory($imagePath, 0755, true);
             }
+
+            $manager = new ImageManager(new Driver());
+            $image = $manager->read($imageFile->getPathname());
+
+            $imageFullPath = $imagePath . $imageName . '.webp';
+            $image->save($imageFullPath);
+
+            $newbanner->image = $imageName . '.webp';
+            $newbanner->image_alt = $imageName;
             
-    
-            // Tag
-            if ($request->tag) {
-                $tags = array_map(fn($item) => ucfirst($item), $request->tag);
-            
-                $tagIds = [];
-                foreach ($tags as $tagName) {
-                    $formattedTagName = Str::title($tagName);
-                    $slug = Str::slug($tagName);
-    
-                    $tag = ArticleTag::firstOrCreate(
-                        ['slug' => $slug],
-                        ['tag' => $formattedTagName]
-                    );
-    
-                    $tagIds[] = $tag->id;
-                }
-    
-                $newarticle->articletag()->attach($tagIds);
-            }
-            
-            // Category
-            if ($request->category) {
-                $category = array_map(fn($item) => ucfirst($item), $request->category);
-            
-                $categoryIds = [];
-                foreach ($category as $categoryName) {
-                    $formattedCategoryName = Str::title($categoryName);
-                    $slug = Str::slug($categoryName);
-    
-                    $category = ArticleCategory::firstOrCreate(
-                        ['slug' => $slug],
-                        ['category' => $formattedCategoryName]
-                    );
-    
-                    $categoryIds[] = $category->id;
-                }
-    
-                $newarticle->articlecategory()->attach($categoryIds);
-            }
-    
-            // Gallery
-            if ($request->has('image_gallery') && !empty($request->image_gallery)) {
-                foreach ($request->image_gallery as $image) {
-                    $newgallery = new ArticleGallery;
-                    $newgallery->article_id = $newarticle->id;
-            
-                    // Pastikan image adalah instance dari UploadedFile
-                    if ($image instanceof \Illuminate\Http\UploadedFile && $image->isValid()) {
-                        // Ambil nama file tanpa ekstensi
-                        $originalName = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
-                        
-                        // Tambahkan tanggal saat ini
-                        $currentDate = now()->format('YmdHis');
-                        
-                        // Gabungkan nama file dan tanggal input
-                        $imageName = $originalName . '_' . $currentDate;
-            
-                        $imagePath = public_path('storage/images/article/gallery/');
-    
-                        if (!File::exists($imagePath)) {
-                            File::makeDirectory($imagePath, 0755, true);
-                        }
-            
-                        $manager = new ImageManager(new Driver());
-                        $imageOptimized = $manager->read($image->getPathname());
-                        $imageFullPath = $imagePath . $imageName . '.webp';
-                        $imageOptimized->save($imageFullPath);
-            
-                        // Simpan nama file dengan ekstensi .webp
-                        $newgallery->image = $imageName . '.webp';
-                        $newgallery->image_alt = $imageName;
-                    }
-            
-                    $newgallery->save();
-                }
-            }
-    
-            // Article Show
-            $newarticleshow = new ArticleShow;
-    
-            if ($request->no_tlp) {
-                $no_tlp = $request->no_tlp;
+            $newbanner->save();
+        }
         
-                if (substr($no_tlp, 0, 1) === '0') {
-                    $no_tlp = '+62' . substr($no_tlp, 1);
-                }
+
+        // Tag
+        if ($request->tag) {
+            $tags = array_map(fn($item) => ucfirst($item), $request->tag);
         
-                $phoneNumber = PhoneNumber::firstOrCreate(
-                    [
-                        'no_tlp' => $no_tlp,
-                        'type' => 'article'    
-                    ]
+            $tagIds = [];
+            foreach ($tags as $tagName) {
+                $formattedTagName = Str::title($tagName);
+                $slug = Str::slug($tagName);
+
+                $tag = ArticleTag::firstOrCreate(
+                    ['slug' => $slug],
+                    ['tag' => $formattedTagName]
                 );
-                
-                $newarticleshow->phone_number_id = $phoneNumber->id;
+
+                $tagIds[] = $tag->id;
+            }
+
+            $newarticle->articletag()->attach($tagIds);
+        }
+        
+        // Category
+        if ($request->category) {
+            $category = array_map(fn($item) => ucfirst($item), $request->category);
+        
+            $categoryIds = [];
+            foreach ($category as $categoryName) {
+                $formattedCategoryName = Str::title($categoryName);
+                $slug = Str::slug($categoryName);
+
+                $category = ArticleCategory::firstOrCreate(
+                    ['slug' => $slug],
+                    ['category' => $formattedCategoryName]
+                );
+
+                $categoryIds[] = $category->id;
+            }
+
+            $newarticle->articlecategory()->attach($categoryIds);
+        }
+
+        // Gallery
+        if ($request->has('image_gallery') && !empty($request->image_gallery)) {
+            foreach ($request->image_gallery as $image) {
+                $newgallery = new ArticleGallery;
+                $newgallery->article_id = $newarticle->id;
+        
+                // Pastikan image adalah instance dari UploadedFile
+                if ($image instanceof \Illuminate\Http\UploadedFile && $image->isValid()) {
+                    // Ambil nama file tanpa ekstensi
+                    $originalName = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
+                    
+                    // Tambahkan tanggal saat ini
+                    $currentDate = now()->format('YmdHis');
+                    
+                    // Gabungkan nama file dan tanggal input
+                    $imageName = $originalName . '_' . $currentDate;
+        
+                    $imagePath = public_path('storage/images/article/gallery/');
+
+                    if (!File::exists($imagePath)) {
+                        File::makeDirectory($imagePath, 0755, true);
+                    }
+        
+                    $manager = new ImageManager(new Driver());
+                    $imageOptimized = $manager->read($image->getPathname());
+                    $imageFullPath = $imagePath . $imageName . '.webp';
+                    $imageOptimized->save($imageFullPath);
+        
+                    // Simpan nama file dengan ekstensi .webp
+                    $newgallery->image = $imageName . '.webp';
+                    $newgallery->image_alt = $imageName;
+                }
+        
+                $newgallery->save();
+            }
+        }
+
+        // Article Show
+        $newarticleshow = new ArticleShow;
+
+        if ($request->no_tlp) {
+            $no_tlp = $request->no_tlp;
+    
+            if (substr($no_tlp, 0, 1) === '0') {
+                $no_tlp = '+62' . substr($no_tlp, 1);
             }
     
-            $newarticleshow->article_id = $newarticle->id;
-            $newarticleshow->banner = $newbanner->image ?? null;
-            $newarticleshow->judul = $newarticle->judul;
-            $newarticleshow->slug = Str::slug($newarticleshow->judul);
-            $newarticleshow->article = $newarticle->article;
-            $newarticleshow->template_id = $request->template_id;
-            $newarticleshow->status = $request->status;
-            $newarticleshow->telephone = $request->tlp;
-            $newarticleshow->whatsapp = $request->wa;
-    
-            if ($request->status === 'schedule') {
-                $newarticleshow->created_at = $request->release;
-            }
-    
-            $newarticleshow->save();
-    
-            // Article Gallery Show
-            $newarticlegallery = ArticleGallery::where('article_id', $newarticle->id)->get();
-    
+            $phoneNumber = PhoneNumber::firstOrCreate(
+                [
+                    'no_tlp' => $no_tlp,
+                    'type' => 'article'    
+                ]
+            );
+            
+            $newarticleshow->phone_number_id = $phoneNumber->id;
+        }
+
+        $newarticleshow->article_id = $newarticle->id;
+        $newarticleshow->banner = $newbanner->image ?? null;
+        $newarticleshow->judul = $newarticle->judul;
+        $newarticleshow->slug = Str::slug($newarticleshow->judul);
+        $newarticleshow->article = $newarticle->article;
+        $newarticleshow->template_id = $request->template_id;
+        $newarticleshow->status = $request->status;
+        $newarticleshow->telephone = $request->tlp;
+        $newarticleshow->whatsapp = $request->wa;
+
+        if ($request->status === 'schedule') {
+            $newarticleshow->created_at = $request->release;
+        }
+
+        $newarticleshow->save();
+
+        // Article Gallery Show
+        $newarticlegallery = ArticleGallery::where('article_id', $newarticle->id)->get();
+
+        if ($newarticlegallery) {
             foreach ($newarticlegallery as $item) {
                 $newgalleryshow = new ArticleShowGallery;
                 
@@ -286,9 +284,9 @@ class ArticleShowController extends Controller
     
                 $newgalleryshow->save();
             }
-    
-            return redirect()->route('article.index')->with('success', 'Artikel berhasil disimpan.');
-        });
+        }
+        
+        return redirect()->route('article.index')->with('success', 'Artikel berhasil disimpan.');
     }
 
     /**
@@ -326,9 +324,7 @@ class ArticleShowController extends Controller
                     Rule::unique('article_shows')->ignore($articleShow->id),
                 ],
                 'category' => 'array',
-                'category.*' => 'max:255',
                 'tag' => 'array',
-                'tag.*' => 'max:255',
                 'article' => 'required',
             ]);
     
@@ -399,10 +395,12 @@ class ArticleShowController extends Controller
         if ($request->hasFile('image')) {
             $banner = ArticleBanner::where('article_id', $articleShow->article_id)->first();
 
-            $path = public_path('storage/images/article/banner/' . $banner->image);
-
-            if (file_exists($path)) {
-                unlink($path);
+            if ($banner) {
+                $path = public_path('storage/images/article/banner/' . $banner->image);
+    
+                if (file_exists($path)) {
+                    unlink($path);
+                }
             }
 
             $imageFile = $request->file('image');
@@ -419,6 +417,11 @@ class ArticleShowController extends Controller
 
             $imageFullPath = $imagePath . $imageName . '.webp';
             $image->save($imageFullPath);
+
+            if (!$banner) {
+                $banner = new ArticleBanner();
+                $banner->article_id = $newarticle->id;
+            }
 
             $banner->image = $imageName . '.webp';
             $banner->image_alt = $imageName;
